Add ABDADA parallel endgame solver
Implement ABDADA (Alpha-Beta Distributed Distributed Algorithm) for
multi-threaded endgame solving with the following features:

- Parallel search with configurable thread count (num_threads parameter)
- Transposition table nproc tracking with optimized 256K entry table
- Relaxed memory ordering for atomic operations to reduce contention
- Stack-allocated deferred move array (MAX_DEFERRED_STACK=64)
- Aspiration windows for iterative deepening
- Early termination via atomic search_complete flag
- Per-ply callback support for PV output during search
- Benchmark test suite for performance testing

Achieves ~24x speedup on deep endgames (e.g., 12-ply solve from 152s
to ~6.3s with 8 threads).

+++ test/heat_map_test.c
@@ -36,8 +36,8 @@ void assert_heat_map_count(const char *name, int row, int col, heat_map_t type,
   if (actual_count != expected_count) {
     printf("heat map counts mismatch for %s:\n", name);
     printf("row: %d, col: %d, type: %d\n", row, col, type);
-    printf("actual:   %lu\n", actual_count);
-    printf("expected: %lu\n", expected_count);
+    printf("actual:   %llu\n", (unsigned long long)actual_count);
+    printf("expected: %llu\n", (unsigned long long)expected_count);
     assert(0);
   }
 }
